# backend/internal/handlers/openapi.yaml
openapi: 3.0.3
info:
  title: RallyUp API
  description: RallyUp Badminton Club management API — member registration, session scheduling, RSVP tracking, and notifications.
  version: 1.0.0
servers:
  - url: /api
    description: API root
...
security:
  - bearerAuth: []
paths:
...
  /admin/users/{id}/invitation:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: ["Admin: Members"]
      summary: Preview an invitation without sending email
      description: Approved admins only. The target must be approved and not yet signed in. The preview includes exact HTML/plain text, send controls, and latest attempt status.
      responses:
        "200":
          description: Invitation preview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPreview" }
        "403":
          description: Approved admin access required
        "409":
          description: Member is not eligible for an invitation
    post:
      tags: ["Admin: Members"]
      summary: Send or resend the reviewed invitation
      description: Obeys both notification stops. Reusing a request ID returns its saved result without another send. Provider acceptance does not confirm inbox delivery.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationSendRequest" }
      responses:
        "200":
          description: Saved attempt, including accepted, failed, pending or unknown results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationDelivery" }
        "409":
          description: Sending paused, stale preview, or ineligible member
        "429":
          description: Wait one minute between sends to this address
  /admin/users/{id}/invitation/test:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      tags: ["Admin: Members"]
      summary: Send a test copy only to the signed-in admin
      description: Requires the separate test-email setting. Ignores caller-supplied destinations. Keeps member invitations and routine notifications paused. Does not mark a member invitation sent.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationSendRequest" }
      responses:
        "200":
          description: Test attempt result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationDelivery" }
        "403":
          description: Approved admin access required
        "409":
          description: Test sending disabled, provider unavailable, or stale preview
        "429":
          description: Wait one minute between tests
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 JWT token
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    InvitationSendRequest:
      type: object
      required: [request_id, expected_email]
      properties:
        request_id: { type: string, format: uuid }
        expected_email: { type: string, format: email, description: Member email shown in the reviewed preview }
    InvitationDelivery:
      type: object
      required: [id, user_id, recipient_email, is_test, status, message, created_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        recipient_email: { type: string, format: email }
        is_test: { type: boolean }
        status: { type: string, enum: [pending, accepted, failed, unknown] }
        message: { type: string }
        accepted_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    InvitationPreview:
      type: object
      properties:
        subject: { type: string }
        html: { type: string }
        text: { type: string }
        login_url: { type: string, format: uri }
        recipient_email: { type: string, format: email }
        from_email: { type: string }
        test_recipient: { type: string, format: email }
        can_send: { type: boolean }
        can_test: { type: boolean }
        send_blocked_reason: { type: string }
        test_blocked_reason: { type: string }
        last_invitation: { $ref: "#/components/schemas/InvitationDelivery" }
        last_test: { $ref: "#/components/schemas/InvitationDelivery" }
...